backend/pyproject.toml
[project]
name = "aisee-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.16",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "numpy>=2.1.1",
    "pydantic[email]>=2.11.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "ultralytics>=8.3.100",
    "uvicorn>=0.34.0",
]
